fix: 添加 autoUpload 属性，给外部手动控制上传用
