+++ packages/zent/src/upload/utils/wrap-promise.ts
@@ -1,0 +1,8 @@
+import isPromise from '../../utils/isPromise';
+
+export function wrapPromise(condition: boolean | Promise<any>) {
+  if (isPromise(condition)) {
+    return condition;
+  }
+  return condition ? Promise.resolve() : Promise.reject();
+}
